Conflicts:
	src/main/java/edu/hm/cs/fs/scriptinat0r7/config/SecurityConfiguration.java
	src/main/webapp/WEB-INF/web.xml

+++ src/test/java/edu/hm/cs/fs/scriptinat0r7/model/ScriptDocumentTest.java
@@ -1,8 +1,6 @@
 package edu.hm.cs.fs.scriptinat0r7.model;
 
-import static org.junit.Assert.assertArrayEquals;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertSame;
+import static org.junit.Assert.*;
 
 import java.util.Collections;
 import java.util.HashSet;
@@ -18,7 +16,7 @@ import edu.hm.cs.fs.scriptinat0r7.model.enums.ReviewState;
  * Test class for the {@code ScriptDocument}.
  */
 public class ScriptDocumentTest {
-    
+
     /**
      * Tests the equals contract for equals() and hashCode().
      */
@@ -35,9 +33,9 @@ public class ScriptDocumentTest {
         ScriptDocument document = new ScriptDocument();
         document.setFile(new byte[] {0x00, 0x01, 0x02, -0x80});
         int expected = 4;
-        
+
         int actual = document.getFileSize();
-        
+
         assertEquals("The size of the file is not the expected.", expected, actual);
     }
 
@@ -48,10 +46,10 @@ public class ScriptDocumentTest {
     public void testGetAndSetFile() {
         ScriptDocument document = new ScriptDocument();
         byte[] expected = new byte[] {0x00, 0x01, 0x02, -0x80};
-        
+
         document.setFile(expected);
         byte[] actual = document.getFile();
-        
+
         assertArrayEquals("The files of the script document are not equal.", expected, actual);
     }
 
@@ -62,13 +60,13 @@ public class ScriptDocumentTest {
     public void testGetAndSetSortnumber() {
         ScriptDocument document = new ScriptDocument();
         int expected = 1;
-        
+
         document.setSortnumber(expected);
         int actual = document.getSortnumber();
-        
+
         assertEquals("The sortnumber of the script document is not equal.", expected, actual);
     }
-    
+
     /**
      * Tests the getter and setter for the review state.
      */
@@ -76,10 +74,10 @@ public class ScriptDocumentTest {
     public void testGetAndSetReviewState() {
         ScriptDocument document = new ScriptDocument();
         ReviewState expected = ReviewState.PROFESSORAPPROVED;
-        
+
         document.setReviewState(expected);
         ReviewState actual = document.getReviewState();
-        
+
         assertSame("The review state of the script document is not the same.", expected, actual);
     }
 
@@ -90,10 +88,10 @@ public class ScriptDocumentTest {
     public void testGetAndSetPassword() {
         ScriptDocument document = new ScriptDocument();
         String expected = "top_secret";
-        
+
         document.setPassword(expected);
         String actual = document.getPassword();
-        
+
         assertSame("The password of the script document is not the same.", expected, actual);
     }
 
@@ -104,10 +102,10 @@ public class ScriptDocumentTest {
     public void testGetAndSetFilename() {
         ScriptDocument document = new ScriptDocument();
         String expected = "datenkommunikation.pdf";
-        
+
         document.setFilename(expected);
         String actual = document.getFilename();
-        
+
         assertSame("The filename of the script document is not the same.", expected, actual);
     }
 
@@ -118,10 +116,10 @@ public class ScriptDocumentTest {
     public void testGetAndSetNote() {
         ScriptDocument document = new ScriptDocument();
         String expected = "some notes...";
-        
+
         document.setNote(expected);
         String actual = document.getNote();
-        
+
         assertSame("The note of the script document is not the same.", expected, actual);
     }
 
@@ -132,13 +130,13 @@ public class ScriptDocumentTest {
     public void testGetAndSetScripts() {
         ScriptDocument document = new ScriptDocument();
         Set<Script> expected = Collections.emptySet();
-        
+
         document.setScripts(expected);
         Set<Script> actual = document.getScripts();
-        
+
         assertSame("The parent scripts of the script document are not the same.", expected, actual);
     }
-    
+
     /**
      * Tests the add method of the scripts.
      */
@@ -151,7 +149,7 @@ public class ScriptDocumentTest {
 
         assertEquals("The set of scripts has not 1 element.", 1, document.getScripts().size());
     }
-    
+
     /**
      * Tests the add method of the scripts when there is already an existing set.
      */
@@ -172,11 +170,11 @@ public class ScriptDocumentTest {
     @Test
     public void testGetAndSetHashvalue() {
         ScriptDocument document = new ScriptDocument();
-        int expected = 123456789;
-        
+        Long expected = 123456789L;
+
         document.setHashvalue(expected);
-        int actual = document.getHashvalue();
-        
+        Long actual = document.getHashvalue();
+
         assertEquals("The hash value of the script document is not the same.", expected, actual);
     }
